+++ src/Command/Connector/ReplaceGedSshWithDepotSftp.php
@@ -105,6 +105,7 @@ class ReplaceGedSshWithDepotSftp extends BaseCommand
         $depotSftpId = $this->connecteurCreationService->createConnecteur(
             'depot-sftp',
             'GED',
+            false,
             $entityId,
             0,
             $label,