+++ template/EntiteMenuGauche.php
@@ -8,22 +8,26 @@
  * @var int $id_e
  * @var bool $permission_on_import_export;
  * @var bool $droitLectureAnnuaire;
+ * @var bool $global
  */
 
-$admninistration_menu["Entite/detail"] = "Informations (entités)";
+$admninistration_menu['Entite/detail'] = 'Informations (entités)';
 
 if ($droitLectureOnUtilisateur) {
-    $admninistration_menu["Entite/utilisateur"] = "Utilisateurs";
+    $admninistration_menu['Entite/utilisateur'] = 'Utilisateurs';
 }
 
 if ($droit_lecture_on_connecteur) {
-    $admninistration_menu["Entite/connecteur"] = "Connecteurs" . ($id_e ? "" : " globaux");
-    $admninistration_menu["Flux/index"] = $id_e ? "Types de dossier (association)" : 'Associations connecteurs globaux';
+    $admninistration_menu['Entite/connecteur?global=0'] = 'Connecteurs' . ($id_e ? '' : " d'entités");
+    if ($id_e === 0) {
+        $admninistration_menu['Entite/connecteur?global=1'] = 'Connecteurs globaux';
+    }
+    $admninistration_menu['Flux/index'] = $id_e ? 'Types de dossier (association)' : 'Associations connecteurs globaux';
 }
 
 if (! empty($permission_on_import_export)) {
-    $admninistration_menu["Entite/exportConfig"] = "Export de la configuration";
-    $admninistration_menu["Entite/importConfig"] = "Import de la configuration";
+    $admninistration_menu['Entite/exportConfig'] = 'Export de la configuration';
+    $admninistration_menu['Entite/importConfig'] = 'Import de la configuration';
 }
 
 if ($droitLectureAnnuaire) {
@@ -47,7 +51,9 @@ $donnees_menu['Entite/agents'] = 'Agents (Actes)';
             <?php foreach ($admninistration_menu as $url => $libelle) : ?>
                 <li>
                     <a class="<?php echo $menu_gauche_select == $url ? "actif" : "" ?>"
-                       href='<?php $this->url(get_hecho($url . "?id_e=$id_e")); ?>'><?php echo $libelle ?></a>
+                       href='<?php $this->url(get_hecho($url . (parse_url($url, PHP_URL_QUERY) ? '&' : '?')) . "id_e=$id_e"); ?>'>
+                        <?php echo $libelle ?>
+                    </a>
                 </li>
             <?php endforeach; ?>
         </ul>